refactor: Session을 가져올때 연관된 Students도 가져오도록 수정

+++ src/main/java/nextstep/students/domain/StudentRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 public interface StudentRepository {
 
-    int save(Student student);
+    long save(Student student);
 
     Optional<Student> findById(Long id);
